improve GlassController, CRUD

+++ src/Controller/GlassController.php
@@ -9,6 +9,7 @@ use App\Form\ApplicationType;
 use App\Repository\ApplicationRepository;
 use App\Repository\StockRepository;
 use App\Repository\UserRepository;
+use App\Repository\PortfolioRepository;
 use App\Service\DealService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
@@ -20,17 +21,17 @@ class GlassController extends AbstractController
     public function __construct(
         private readonly StockRepository $stockRepository,
         private readonly UserRepository $userRepository,
+        private readonly PortfolioRepository $portfolioRepository,
         private readonly ApplicationRepository $applicationRepository,
         private readonly DealService $dealService
     ) {
 
     }
 
-    #[Route('/glass/stock/{stockId}', methods: ['GET'],  name: 'app_stock_glass')]
+    #[Route('glass/{stockId}', methods: ['GET'],  name: 'app_stock_glass')]
     public function getStockGlass(int $stockId): Response
     {
         
-        //$this->stockRepository->findBy(['id' => $stockId]);
         $stock = $this->stockRepository->findById($stockId);
         if ($stock === null){
             throw $this->createNotFoundException("Stock not found");
@@ -43,16 +44,32 @@ class GlassController extends AbstractController
         ]);
     }
 
-    #[Route('/glass/stock/{stockId}', methods: ['POST'], name: 'app_stock_glass_create_application')]
-    public function createApplication(int $stockId, Request $request): Response
+    #[Route('application/create', methods: ['POST'], name: 'app_stock_glass_create_application')]
+    public function createApplication(Request $request): Response
     {
         $userId = $request->getPayload()->get('user_id');
+        $stockId = $request->getPayload()->get('stock_id');
         $quantity = $request->getPayload()->get('quantity');
         $price = $request->getPayload()->get('price');
         $action = ActionEnum::from($request->getPayload()->get('action'));
+        $portfolioId = $request->getPayload()->get('portfolio_id');
 
         $stock = $this->stockRepository->findById($stockId);
         $users = $this->userRepository->findBy(['id' => $userId]);
+        $portfolio = $this->portfolioRepository->find($portfolioId);
+
+        if ($stock === null) {
+            throw $this->createNotFoundException("Stock not found");
+        }
+
+        $user = $this->userRepository->find($userId);
+        if ($user === null) {
+            throw $this->createNotFoundException("User not found");
+        }
+
+        if ($portfolio === null) {
+            throw $this->createNotFoundException("Portfolio not found");
+        }
 
         $application = new Application;
         $application->setStock($stock);
@@ -60,6 +77,7 @@ class GlassController extends AbstractController
         $application->setAction($action);
         $application->setPrice($price);
         $application->setUser(current($users));
+        $application->setPortfolio($portfolio);
 
         $appropriateApplication = $this->applicationRepository->findAppropriate($application);
         if ($appropriateApplication !== null){
@@ -72,17 +90,46 @@ class GlassController extends AbstractController
     }
 
 
-    #[Route('/glass/stock/{stockId}', name: 'app_stock_glass_update_application', methods: ['PATCH'])]
+    #[Route('application/update', name: 'app_stock_glass_update_application', methods: ['PUT'])]
 
-    public function updateApplication(int $stockId, Request $request): Response
+    public function updateApplication(Request $request): Response
     {
         $applicationId = $request->getPayload()->get('application_id');
+        $userId = $request->getPayload()->get('user_id');
+        $stockId = $request->getPayload()->get('stock_id');
         $quantity = $request->getPayload()->get('quantity');
         $price = $request->getPayload()->get('price');
+        $action = ActionEnum::from($request->getPayload()->get('action'));
+        $portfolioId = $request->getPayload()->get('portfolio_id');
 
+        $stock = $this->stockRepository->findById($stockId);
+        $users = $this->userRepository->findBy(['id' => $userId]);
+        $portfolio = $this->portfolioRepository->find($portfolioId);
         $application = $this->applicationRepository->find($applicationId);
+
+        if ($application === null){
+            throw $this->createNotFoundException("Application not found");
+        }
+
+        if ($stock === null) {
+            throw $this->createNotFoundException("Stock not found");
+        }
+
+        $user = $this->userRepository->find($userId);
+        if ($user === null) {
+            throw $this->createNotFoundException("User not found");
+        }
+
+        if ($portfolio === null) {
+            throw $this->createNotFoundException("Portfolio not found");
+        }
+
+        $application->setStock($stock);
         $application->setQuantity($quantity);
+        $application->setAction($action);
         $application->setPrice($price);
+        $application->setUser(current($users));
+        $application->setPortfolio($portfolio);
 
         $appropriateApplication = $this->applicationRepository->findAppropriate($application);
         if($appropriateApplication !== null){
@@ -95,11 +142,15 @@ class GlassController extends AbstractController
         return new Response("OK", Response::HTTP_ACCEPTED);
     }
 
-    #[Route('/glass/stock/{stockId}', name: 'app_stock_glass_delete_application', methods: ['DELETE'])]
-    public function deleteApplication(int $stockId, Request $request): Response
+    #[Route('application/delete', name: 'app_stock_glass_delete_application', methods: ['DELETE'])]
+    public function deleteApplication(Request $request): Response
     {
         $applicationId = $request->getPayload()->get('application_id');
         $application = $this->applicationRepository->find($applicationId);
+
+        if ($application === null){
+            throw $this->createNotFoundException("Application not found");
+        }
 
         $this->applicationRepository->removeApplication($application);
